+++ src/pages/Dashboard/Payment/TicketPage.js
@@ -1,0 +1,3 @@
+export default function TicketPage() {
+  return <></>;
+}
